Create GameKeyboard.java
This class basically analyzes input. Based on the keys we can move our character, or do anything we to INSIDE AND OUTSIDE of the GameKeyboard object.

This is because there are TWO processInput() methods. One that will execute instructions from within the object, and one that can return a boolean value based on the argument passed to it (a KeyCode). It will return true if the key is currently being pressed, and false if not.