Create AStar.java
This implementation has not yet been adjusted for however we represent the map.  It currently works on discrete points.  If you would like to try it, I will upload a simple MazeReader class and some mazes for it to navigate.  I'll try to get it working on our map, and to get enemies chasing the player.