Added buildPath and step methods
buildPath() will build a path through the given terrain, using either a simple, direct path, or an AStar path.  Step will make the Actor take one step along this path.  If step() runs out of points, it sets the path to null, which allows a new path to be built.