Added audio support!
So I just edited the class by adding to the datafields and the creategui method to start looping the ambience .wav file, a subsequently adding a loopclip method.

It worked for me, so here's hoping that I didn't screw anything up. Let me know if anything looks weird or if I should change anything!